requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.10
python-dotenv==0.19.0
requests==2.26.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
pytest==6.2.5
coverage==5.5
python-dateutil==2.8.2
six==1.16.0
urllib3==1.26.6
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
attrs==21.2.0
iniconfig==1.1.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
toml==0.10.2
Flask-Login==0.6.2
supabase==2.3.1
# Google Calendar Integration
google-api-python-client==2.87.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
flask-bootstrap==3.3.7.1
# Email and Document Handling
email-validator==2.0.0
pypdf2==3.0.1
pillow==9.5.0
xlrd==2.0.1
openpyxl==3.1.2
# Testing Libraries
pytest-flask==1.2.0
pytest-cov==4.1.0
pytest-mock==3.11.1
factory-boy==3.3.0 